refactor: add shared chart payload helper
