+++ codexhorary/frontend/package.json
@@ -30,7 +30,8 @@
     "fix-and-package": "node scripts/fix-and-package.js",
     "build:appx-from-working": "npm run build && npm run build-backend-exe && npm run electron:pack && npm run fix-and-package",
     "validate:icons": "node validate-icons.js",
-    "prebuild": "npm run validate:icons"
+    "prebuild": "npm run validate:icons",
+    "test": "node tests/buildChartPayload.test.mjs"
   },
   "dependencies": {
     "lucide-react": "^0.263.1",
